// ui/src/app/model/model.component.ts
export class Vehicle {
    
    uniqueId: string = '';
    brand: string = '';
    model: string = '';
    licensePlate: string = '';
    fuelConsumption: number;
    
    constructor() {}
}

// #################
// # GOOGLE MAPS API
// #################
export class GoogleApiDuration {
   
    text: String = '';
    value: number = 0;
    
    constructor() {}
}

export class GoogleApiDistance {
    
    text: String = '';
    value: number = 0;
    
    constructor() {}
}
// #################